refactor(pools): reorganize pools classes
