+++ test/test-util.ts
@@ -1,7 +1,6 @@
 import { ethers } from 'ethers'
 
-import { Erc20, Pools } from '../src'
-import { SelfProvider } from '../src/pools/self-provider'
+import { Erc20, Pools, SelfProvider } from '../src'
 import { CollateralToken } from '../src/types/pools'
 
 export const ERC20_COLLATERAL_POOL_ETH_ADDRESS =
